pies/pi2/pyproject.toml
[project]
name = "pi2"
version = "2026.2.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "gpiohero",
    "paho-mqtt>=2.1.0",
]

[tool.ruff]
target-version = "py39"
extend = "../pyproject.toml"
